pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["twin_core", "mcp_core", "skill_registry", "tool_registry", "domain_agents", "orchestrator", "api_gateway", "observability", "digital_twin"]

[project]
name = "metaforge"
version = "0.1.0"
description = "Local-first control plane for hardware design"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "pydantic-ai>=0.1",
    "structlog>=24.0",
    "fastapi>=0.100",
    "uvicorn[standard]>=0.30",
    "python-multipart>=0.0.9",
    "httpx>=0.27",
    "pyyaml>=6.0",
    "opentelemetry-api>=1.20",
    "opentelemetry-sdk>=1.20",
    "opentelemetry-exporter-otlp-proto-grpc>=1.20",
    "opentelemetry-instrumentation-fastapi>=0.43b0",
]

[project.scripts]
forge-server = "api_gateway.server:main"

[project.optional-dependencies]
neo4j = [
    "neo4j>=5.0",
]
kafka = [
    "aiokafka>=0.9",
]
assistant = [
    "watchfiles>=0.21",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
    "mypy>=1.13",
    "httpx>=0.27",
]
temporal = [
    "temporalio>=1.5",
]

knowledge = [
    "asyncpg>=0.29",
    "pgvector>=0.2",
    "sentence-transformers>=2.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.11"
plugins = ["pydantic.mypy"]
warn_unused_configs = true
check_untyped_defs = true
exclude = [
    "cli/",
    "ide_assistants/",
    "dashboard/",
    "examples/",
    "tools/",
    "docs/",
    "venv/",
    "build/",
    "dist/",
]
ignore_missing_imports = true
